extract type checking to separate class
